Blog Details Page Changes:
Added a new path in the blog folder specifically for the blog details page.
Updated the start position in the animation to keep the circle centered on 'blog' even when transitioning to 'blog/id'.
Modified the router-dom configuration to navigate to the blog details page without altering the header or footer components.
Implemented the initial section on the blog details page.